ai-service/requirements.txt
# CrewAI and AI Dependencies
crewai==0.70.1
crewai-tools==0.12.1
google-generativeai==0.8.3
langchain-google-genai==2.0.5

# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# Database and Utilities
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0
httpx==0.27.2

# Additional AI Tools
langchain>=0.2.16,<0.3.0
langchain-community>=0.2.16,<0.3.0
